eduture_rl/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
numpy==2.2.1
scipy==1.14.1
python-multipart==0.0.20
email-validator==2.2.0
